updated test, refactored moderator
