+++ lib/cluster/moderator/index.js
@@ -2,7 +2,6 @@
 
 var messageModule = require('./../services/messaging');
 var elasticError = require('../../utils/error_utils').elasticError;
-var _ = require('lodash');
 
 module.exports = function(context) {
     var logger = context.foundation.makeLogger('moderator', 'moderator', {module: 'moderator'});
