+++ lib/cluster/moderator/modules/elasticsearch.js
@@ -3,8 +3,6 @@ var getClient = require('../../../utils/config').getClient;
 var Promise = require('bluebird');
 var elasticError = require('../../../utils/error_utils').elasticError;
 var _ = require('lodash');
-var nodeInfo = require('../../../data_sources/elasticsearch').nodeInfo;
-var nodeStats = require('../../../data_sources/elasticsearch').nodeStats;
 
 module.exports = function(context, logger) {
     var limit = context.sysconfig.teraslice.moderator_limit;
@@ -16,10 +14,12 @@ module.exports = function(context, logger) {
     var clients = [];
 
     var state = {};
+    //used to attach the elasticsearch module for each client name
+    var clientMapping = {};
 
     //used to set state and queue size limits
     function initializeConnection(client) {
-        return nodeInfo(client)
+        return client.__esModule.nodeInfo()
             .then(function(results) {
                 let client_name = client.__esConnection;
                 //top level connection name
@@ -61,6 +61,7 @@ module.exports = function(context, logger) {
                 _.forOwn(context.sysconfig.terafoundation.connectors.elasticsearch, function(val, key) {
                     let client = getClient(context, {connection: key}, 'elasticsearch');
                     client.__esConnection = key;
+                    client.__esModule =  require('../../../data_sources/elasticsearch')(client, logger, null);
 
                     initializeConnection(client)
                         .then(function() {
@@ -128,7 +129,7 @@ module.exports = function(context, logger) {
         let resumeJobs = [];
 
         return Promise.map(clients, function(client) {
-            return nodeStats(client)
+            return client.__esModule.nodeStats()
                 .then(function(results) {
                     let client_name = client.__esConnection;
                     _.forOwn(results.nodes, function(stats, node_name) {
