+++ lib/cluster/runners/job.js
@@ -1,9 +1,6 @@
 'use strict';
 
-var fs = require('fs');
 var Promise = require('bluebird');
-
-var exceptions = require('../../utils/exceptions');
 var insertAnalyzers = require('../../utils/analytics').insertAnalyzers;
 
 /*
