+++ lib/cluster/runners/op.js
@@ -1,7 +1,6 @@
 'use strict';
 
 var path = require('path');
-var exceptions = require('../../utils/exceptions');
 var existsSync = require('../../utils/file_utils').existsSync;
 var fs = require('fs');
 var path = require('path');
