+++ lib/cluster/services/api.js
@@ -4,7 +4,6 @@ var _ = require('lodash');
 var Promise = require('bluebird');
 var bodyParser = require('body-parser');
 var elasticError = require('../../utils/error_utils').elasticError;
-var exceptions = require('../../utils/exceptions');
 var Table = require('easy-table');
