+++ lib/cluster/slicer.js
@@ -20,10 +20,10 @@ var messageModule = require('./services/messaging');
 
 module.exports = function(context) {
     var state_store;
+    var ex_id = process.env.ex_id;
 
     var logger = context.foundation.makeLogger('slicer', 'slicer', {ex_id: ex_id, module: 'slicer'});
     var messaging = messageModule(context, logger);
-    var ex_id = ex_id;
 
     //Stateful variables
     var start = moment();
