+++ lib/cluster/storage/backends/elasticsearch_store.js
@@ -245,9 +245,10 @@ module.exports = function(context, index_name, record_type, id_field, bulk_size)
                         .then(function(results) {
                             return results;
                         })
-                        .error(function(err) {
+                        .catch(function(err) {
                             // It's not really an error if it's just that the index is already there.
-                            if ((err.message.indexOf('IndexAlreadyExistsException') < 0) && (err.message.indexOf('index_already_exists_exception') < 0)) {
+                            //console.log('whats error coming out of here on createIndex',err, typeof err, );
+                            if (err.match(/index_already_exists_exception/) === null) {
                                 var errMsg = elasticError(err);
                                 logger.error(`Could not create index: ${index_name}, error: ${errMsg}`);
                                 return Promise.reject(`Could not create job index, error: ${errMsg}`)
