+++ lib/data_sources/elasticsearch.js
@@ -3,15 +3,11 @@
 var uuid = require('uuid');
 var Promise = require('bluebird');
 var _ = require('lodash');
-var moment = require('moment');
-var dateMath = require('datemath-parser');
 var elasticError = require('../utils/error_utils').elasticError;
 
 // Module to manage persistence in Elasticsearch.
 // All functions in this module return promises that must be resolved to get the final result.
 
-
-// TODO fix autodates
 // TODO elasticError fix, if it needs to be a module
 
 
@@ -243,20 +239,6 @@ module.exports = function(client, logger, opConfig) {
         });
     }
 
-    function _parseDate(date) {
-        var result;
-
-        if (moment(new Date(date)).isValid()) {
-            result = moment(new Date(date));
-        }
-        else {
-            var ms = dateMath.parse(date);
-            result = moment(ms)
-        }
-
-        return result;
-    }
-
     function _errorHandler(retryTimer, fn, reject, logger) {
         return function(err) {
             if (_.get(err, 'body.error.type') === 'es_rejected_execution_exception') {
@@ -278,40 +260,6 @@ module.exports = function(client, logger, opConfig) {
                 reject(errMsg)
             }
         }
-    }
-
-// slicer get dates
-
-    function autoDates(query, givenDate, order, dateFormat) {
-        return client.search(query)
-            .then(function(data) {
-                if (data.hits.hits.length === 0) {
-                    return null
-                }
-
-                if (data.hits.hits[0] && data.hits.hits[0]._source[opConfig.date_field_name] === undefined) {
-                    return Promise.reject(`date_field_name: "${opConfig.date_field_name}" for index: ${opConfig.index} does not exist`);
-                }
-
-                if (givenDate) {
-                    return givenDate;
-                }
-
-                if (order === 'start') {
-                    return _parseDate(data.hits.hits[0]._source[opConfig.date_field_name])
-                }
-                else {
-                    //end date is non-inclusive, adding 1s so range will cover it
-                    var date = data.hits.hits[0]._source[opConfig.date_field_name];
-                    var time = moment(date).add(1, 's');
-                    return _parseDate(time.format(dateFormat));
-                }
-            })
-            .catch(function(err) {
-                var errMsg = elasticError(err);
-                logger.error(errMsg);
-                return Promise.reject(errMsg);
-            });
     }
 
     function _checkVersion(str) {
@@ -407,11 +355,11 @@ module.exports = function(client, logger, opConfig) {
         });
     }
 
-    function nodeInfo(client) {
+    function nodeInfo() {
         return client.nodes.info();
     }
 
-    function nodeStats(client) {
+    function nodeStats() {
         return client.nodes.stats()
     }
 
@@ -545,7 +493,6 @@ module.exports = function(client, logger, opConfig) {
         create: create,
         update: update,
         remove: remove,
-        autoDates: autoDates,
         version: version,
         putTemplate: putTemplate,
         bulkSend: bulkSend,
