+++ lib/readers/elasticsearch_date_range/slicer.js
@@ -3,6 +3,7 @@
 var Promise = require('bluebird');
 var moment = require('moment');
 var _ = require('lodash');
+var dateMath = require('datemath-parser');
 var dateOptions = require('./../../utils/date_utils').dateOptions;
 var dateFormatMS = require('./../../utils/date_utils').dateFormat;
 var dateFormatS = require('./../../utils/date_utils').dateFormatSeconds;
@@ -68,6 +69,20 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         return interval;
     }
 
+    function _parseDate(date) {
+        var result;
+
+        if (moment(new Date(date)).isValid()) {
+            result = moment(new Date(date));
+        }
+        else {
+            var ms = dateMath.parse(date);
+            result = moment(ms)
+        }
+
+        return result;
+    }
+
     function getIndexDate(opConfig, client, date, order) {
 
         var sortObj = {};
@@ -75,7 +90,7 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         var query = null;
 
         if (date) {
-            givenDate = elasticsearch.parseDate(date);
+            givenDate = _parseDate(date);
             query = elasticsearch.buildQuery(opConfig, {count: 1, start: opConfig.start, end: opConfig.end})
         }
         else {
@@ -99,7 +114,32 @@ function newSlicer(context, opConfig, job, retryData, logger, client) {
         }
 
         //using this query to catch potential errors even if a date is given already
-        return elasticsearch.autoDates(query, givenDate, order, dateFormat);
+        return elasticsearch.search(query)
+            .then(function(results) {
+                var data = (results.hits && results.hits.hits) ? results.hits.hits[0]._source : results[0];
+
+                if (data === undefined) {
+                    return null
+                }
+
+                if (data[opConfig.date_field_name] === undefined) {
+                    return Promise.reject(`date_field_name: "${opConfig.date_field_name}" for index: ${opConfig.index} does not exist`);
+                }
+
+                if (givenDate) {
+                    return givenDate;
+                }
+
+                if (order === 'start') {
+                    return _parseDate(data[opConfig.date_field_name])
+                }
+                else {
+                    //end date is non-inclusive, adding 1s so range will cover it
+                    var date = data[opConfig.date_field_name];
+                    var time = moment(date).add(1, 's');
+                    return _parseDate(time.format(dateFormat));
+                }
+            });
     }
 
     function updateJob(opConfig, dates) {
